another bunch of new implicit conversions

+++ include/simgrid/host.h
@@ -64,7 +64,7 @@ XBT_PUBLIC void sg_host_get_disks(const_sg_host_t host, unsigned int* disk_count
 /** @brief Return the speed of the processor (in flop/s), regardless of the current load on the machine. */
 XBT_PUBLIC double sg_host_get_speed(const_sg_host_t host);
 XBT_ATTRIB_DEPRECATED_v330("Please use sg_host_get_speed()") XBT_PUBLIC double sg_host_speed(const_sg_host_t host);
-XBT_PUBLIC double sg_host_get_pstate_speed(const_sg_host_t host, int pstate_index);
+XBT_PUBLIC double sg_host_get_pstate_speed(const_sg_host_t host, unsigned long pstate_index);
 
 XBT_PUBLIC double sg_host_get_available_speed(const_sg_host_t host);
